resumeai/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.2.0
python-dotenv==1.0.1
httpx==0.26.0
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.1.0
google-auth==2.28.1
requests==2.31.0
openai==1.12.0
anthropic==0.18.1
Pillow==10.3.0
aiofiles==23.2.1
email-validator==2.1.0
